requirements.txt
streamlit
pyyaml
groq